pyproject.toml
[project]
name = "py"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0.2",
    "icecream>=2.1.4",
    "jinja2>=3.1.6",
    "inotify>=0.2.10",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
